feat: add solution to modules-task
